updated examples to match ablate refactor
