+++ ablateLibraryClient.cpp
@@ -1,16 +1,18 @@
 #include <environment/runEnvironment.hpp>
+#include <flow/boundaryConditions/essential.hpp>
 #include <mathFunctions/functionFactory.hpp>
 #include <memory>
-#include <mesh/boxMesh.hpp>
-#include <monitors/flow/hdf5OutputFlow.hpp>
-#include <parameters/mapParameters.hpp>
 #include "builder.hpp"
 #include "flow/flow.hpp"
 #include "flow/incompressibleFlow.hpp"
+#include "mesh/boxMesh.hpp"
+#include "monitors/hdf5Monitor.hpp"
+#include "monitors/timeStepMonitor.hpp"
+#include "parameters/mapParameters.hpp"
 #include "solve/timeStepper.hpp"
 #include "utilities/petscOptions.hpp"
 
-int main(int argc, char **argv) {
+int main(int argc, char** argv) {
     // initialize petsc and mpi
     PetscErrorCode ierr = PetscInitialize(&argc, &argv, NULL, NULL);
     CHKERRABORT(PETSC_COMM_WORLD, ierr);
@@ -21,7 +23,7 @@ int main(int argc, char **argv) {
         ablate::environment::RunEnvironment::Setup(runEnvironmentParameters);
 
         // setup any global arguments
-        ablate::utilities::PetscOptions::Set({{"dm_plex_separate_marker", ""}});
+        ablate::utilities::PetscOptionsUtils::Set({{"dm_plex_separate_marker", ""}, {"vel_petscspace_degree", "3"}, {"pres_petscspace_degree", "2"}, {"temp_petscspace_degree", "2"}});
 
         // create a time stepper
         auto timeStepper = ablate::solve::TimeStepper("timeStepper",
@@ -43,13 +45,14 @@ int main(int argc, char **argv) {
                                                        {"fieldsplit_pressure_pc_type", "jacobi"}});
 
         auto mesh = std::make_shared<ablate::mesh::BoxMesh>("simpleMesh",
-                                                            std::map<std::string, std::string>{{"dm_refine", "0"},
-                                                                                               {"vel_petscspace_degree", "3"},
-                                                                                               {"pres_petscspace_degree", "2"},
-                                                                                               {"temp_petscspace_degree", "2"}},
                                                             std::vector<int>{2, 3},
                                                             std::vector<double>{.1, .1},
-                                                            std::vector<double>{.2, .2});
+                                                            std::vector<double>{.2, .2},
+                                                            std::vector<std::string>{} /*boundary*/,
+                                                            true /*simplex*/,
+                                                            std::make_shared<ablate::parameters::MapParameters>(std::map<std::string, std::string>{
+                                                                {"dm_refine", "0"},
+                                                            }));
 
         // setup a flow parameters
         auto parameters = std::make_shared<ablate::parameters::MapParameters>(std::map<std::string, std::string>{{"strouhal", "1.0"},
@@ -69,37 +72,40 @@ int main(int argc, char **argv) {
         auto flow = std::make_shared<ablate::flow::IncompressibleFlow>(
             "FlowField",
             mesh,
-            std::map<std::string, std::string>{},
             parameters,
-            std::vector<std::shared_ptr<ablate::flow::FlowFieldSolution>>{
-                std::make_shared<ablate::flow::FlowFieldSolution>(
-                    "velocity", ablate::mathFunctions::Create("t + x^2 + y^2, t + 2*x^2 - x*y"), ablate::mathFunctions::Create("1.0, 1.0")),
-                std::make_shared<ablate::flow::FlowFieldSolution>(
-                    "pressure", ablate::mathFunctions::Create("x + y - 1"), ablate::mathFunctions::Create("1.0")),
-                std::make_shared<ablate::flow::FlowFieldSolution>(
-                    "temperature", ablate::mathFunctions::Create("t + x + y"), ablate::mathFunctions::Create("1.0")),
+            /* petsc options*/
+            nullptr,
+            /* init functions */
+            std::vector<std::shared_ptr<ablate::mathFunctions::FieldSolution>>{
+                std::make_shared<ablate::mathFunctions::FieldSolution>("velocity", ablate::mathFunctions::Create("t + x^2 + y^2, t + 2*x^2 - x*y"), ablate::mathFunctions::Create("1.0, 1.0")),
+                std::make_shared<ablate::mathFunctions::FieldSolution>("pressure", ablate::mathFunctions::Create("x + y - 1"), ablate::mathFunctions::Create("1.0")),
+                std::make_shared<ablate::mathFunctions::FieldSolution>("temperature", ablate::mathFunctions::Create("t + x + y"), ablate::mathFunctions::Create("1.0")),
             },
-            std::vector<std::shared_ptr<ablate::flow::BoundaryCondition>>{
-                std::make_shared<ablate::flow::BoundaryCondition>(
+            /* boundary conditions*/
+            std::vector<std::shared_ptr<ablate::flow::boundaryConditions::BoundaryCondition>>{
+                std::make_shared<ablate::flow::boundaryConditions::Essential>(
                     "velocity",
                     "wall velocity",
                     "marker",
-                    1,
-                    ablate::mathFunctions::Create([](PetscInt dim, PetscReal time, const PetscReal x[], PetscInt Nf, PetscScalar* u, auto ctx ){
-                        u[0] = time + x[0]*x[0] + x[1]*x[1];
-                        u[1] = time + 2*x[0]*x[0] - x[0]*x[1];
+                    std::vector<int>{1},
+                    ablate::mathFunctions::Create([](PetscInt dim, PetscReal time, const PetscReal x[], PetscInt Nf, PetscScalar* u, auto ctx) {
+                        u[0] = time + x[0] * x[0] + x[1] * x[1];
+                        u[1] = time + 2 * x[0] * x[0] - x[0] * x[1];
                         return 0;
-                    }),/**example showing lambda input**/
+                    }), /**example showing lambda input**/
                     ablate::mathFunctions::Create("1.0, 1.0")),
-                std::make_shared<ablate::flow::BoundaryCondition>(
-                    "temperature", "wall temp", "marker", 1, ablate::mathFunctions::Create("t + x + y + z"), ablate::mathFunctions::Create("1.0"))},
-            std::vector<std::shared_ptr<ablate::flow::FlowFieldSolution>>{});
+                std::make_shared<ablate::flow::boundaryConditions::Essential>(
+                    "temperature", "wall temp", "marker", 1, ablate::mathFunctions::Create("t + x + y + z"), ablate::mathFunctions::Create("1.0"))});
 
         // assume one flow field right now
         flow->SetupSolve(timeStepper.GetTS());
 
-        // setup a monitor
-        auto monitor = std::make_shared<ablate::monitors::flow::Hdf5OutputFlow>();
+        // setup monitors
+        auto hdf5Monitor = std::make_shared<ablate::monitors::Hdf5Monitor>();
+        hdf5Monitor->Register(flow);
+        timeStepper.AddMonitor(hdf5Monitor);
+
+        auto monitor = std::make_shared<ablate::monitors::TimeStepMonitor>();
         monitor->Register(flow);
         timeStepper.AddMonitor(monitor);
